backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
dotenv==0.9.9
flatbuffers==25.2.10
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
onnxruntime==1.22.0
openai==1.68.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.1.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMatting==1.1.13
pyreadline3==3.5.4
python-dotenv==1.0.1
referencing==0.36.2
rembg==2.0.66
requests==2.32.3
rpds-py==0.25.0
scikit-image==0.25.2
scipy==1.15.3
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
